+++ src/app/app-routing.module.ts
@@ -12,6 +12,7 @@ import { ProfileComponent } from './components/profile/profile.component';
 import { HomePageComponent } from './components/home/home-page/home-page.component';
 import { TasksComponent } from './tasks/tasks.component';
 import { NotFoundComponent } from './components/error-handling/not-found/not-found.component';
+import { ChatFeedComponent } from './components/chat-feed/chat-feed.component';
 
 const routes: Routes = [
   { path: '', redirectTo: 'login', pathMatch: 'full' },
@@ -20,11 +21,12 @@ const routes: Routes = [
   { path: 'signup', component: RegisterComponent },
   { path: 'profile', component: ProfileComponent },
   { path: 'home', component: HomePageComponent },
-  { path: 'chatroom', component: ChatroomComponent },
+  { path: 'chats', component: ChatroomComponent },
   { path: 'password_reset', component: ForgotPasswordComponent },
   { path: 'tasks', component: TasksComponent },
   { path: 'add-room',component:AddRoomComponent},
-  { path: 'roomlists',component:RoomlistsComponent},
+  { path: 'roomlist',component:RoomlistsComponent},
+  { path: 'chatroom/:roomname',component:ChatFeedComponent},
   { path: '**', component: NotFoundComponent },
 
 ];
