+++ src/app/components/chat-feed/chat-feed.component.css
@@ -1,9 +1,157 @@
-.feed{
-    display: flex;
-    flex-direction: column;
+
+.example-container {
+  display: flex;
+  padding: 10px;
 }
 
-.message{
-    flex-direction: column;
-    margin: 10px 0;
+.drawer-container {
+  position: absolute;
+  right: 0;
+  left: 0;
+  bottom: 0;
+  top: 0;
+}
+
+.left-drawer {
+  width: 200px;
+}
+
+.users-pane {
+  height: 500px;
+}
+
+.users-card {
+  margin: 5px 20px;
+}
+
+.username {
+  position: absolute;
+  margin-top: 5px;
+  margin-left: 10px;
+}
+
+footer.sticky-footer {
+  position: fixed;
+  bottom: 0;
+  left: 0;
+  width: 100%;
+  padding: 10px;
+  background-color: #ffffff;
+  border-top: solid 1px #efefef;
+}
+
+.message-form {
+  margin-left: 200px;
+}
+
+.message-form-field {
+  width: 94%;
+  margin: 0 4% 0 2%;
+}
+
+.message-box {
+  float: left;
+  width: 98%;
+  margin: 5px 0 0 2%;
+}
+
+.chat-message {
+  width: 80%;
+  min-height: 40px;
+}
+
+.chat-message .right-bubble {
+  position: relative;
+  background: #dcf8c6;
+  border-top-left-radius: .4em;
+  border-bottom-left-radius: .4em;
+  border-bottom-right-radius: .4em;
+  padding: 5px 10px 10px;
+  left: 15%;
+}
+
+.chat-message .right-bubble span.msg-name {
+  font-size: 12px;
+  font-weight: bold;
+  color: green;
+}
+.chat-message .right-bubble span.msg-date {
+  font-size: 10px;
+}
+
+.chat-message .right-bubble:after {
+  content: '';
+  position: absolute;
+  right: 0;
+  top: 13px;
+  width: 0;
+  height: 0;
+  border: 27px solid transparent;
+  border-left-color: #dcf8c6;
+  border-right: 0;
+  border-top: 0;
+  margin-top: -13.5px;
+  margin-right: -27px;
+}
+.chat-message .left-bubble {
+  position: relative;
+  background: lightblue;
+  border-top-right-radius: .4em;
+  border-bottom-left-radius: .4em;
+  border-bottom-right-radius: .4em;
+  padding: 5px 10px 10px;
+  left: 5%;
+}
+.chat-message .left-bubble span.msg-name {
+  font-size: 12px;
+  font-weight: bold;
+  color: blue;
+}
+.chat-message .left-bubble span.msg-date {
+  font-size: 10px;
+}
+.chat-message .left-bubble:after {
+  content: '';
+  position: absolute;
+  left: 0;
+  top: 13px;
+  width: 0;
+  height: 0;
+  border: 27px solid transparent;
+  border-right-color: lightblue;
+  border-left: 0;
+  border-top: 0;
+  margin-top: -13.5px;
+  margin-left: -27px;
+}
+
+.chat-message .chat-status {
+  min-height: 49px;
+}
+
+.chat-message .chat-status .chat-date {
+  display: block;
+  font-size: 10px;
+  font-style: italic;
+  color: #fff;
+  text-shadow: 0px -1px 0px #222, 0px 1px 0px #aaa;
+  height: 15px;
+  left: 10%;
+  right:10%;
+}
+
+.chat-message .chat-status .chat-content-center {
+  padding: 5px 10px;
+  background-color: #e1e1f7;
+  border-radius: 6px;
+  font-size: 12px;
+  color: #555;
+  height: 34px;
+  left: 10%;
+  right:10%;
+}
+
+.chat-content {
+  overflow-y: scroll;
+  height: 600px;
 }
