+++ src/app/components/chat-feed/chat-feed.component.html
@@ -1,5 +1,45 @@
-<div class="feed">
-  <div *ngFor="let message of feed | async" class="message">
-    <app-message [chatMessage]=message></app-message>
-  </div>
+<div class="example-container mat-elevation-z8">
+  <mat-drawer-container class="drawer-container">
+    <mat-drawer mode="side" opened class="left-drawer">
+      <div class="users-pane">
+        <mat-card class="users-card" *ngFor="let user of users">
+          <mat-icon>person</mat-icon> <span class="username">{{user.chatname}}</span>
+        </mat-card>
+      </div>
+    </mat-drawer>
+    <mat-drawer-content class="chat-pane">
+      <div #chatcontent [scrollTop]="scrolltop" class="chat-content">
+        <div class="message-box" *ngFor="let chat of chats">
+          <div class="chat-status" text-center *ngIf="chat.type==='join'||chat.type==='exit';else message">
+            <span class="chat-date">{{chat.date | date:'short'}}</span>
+            <span class="chat-content-center">{{chat.message}}</span>
+          </div>
+          <ng-template #message>
+            <div class="chat-message">
+              <div class="right-bubble"
+                [ngClass]="{'right-bubble': chat.chatname === chatname, 'left-bubble': chat.chatname !== chatname}">
+                <span class="msg-name" *ngIf="chat.chatname === chatname">Me</span>
+                <span class="msg-name" *ngIf="chat.chatname !== chatname">{{chat.chatname}}</span>
+                <span class="msg-date"> at {{chat.date | date:'short'}}</span>
+                <p text-wrap>{{chat.message}}</p>
+              </div>
+            </div>
+          </ng-template>
+        </div>
+      </div>
+      <footer class="sticky-footer">
+        <form class="message-form" [formGroup]="chatForm" (ngSubmit)="onFormSubmit(chatForm.value)">
+          <mat-form-field class="message-form-field">
+            <input matInput placeholder="Enter message here" formControlName="message" [errorStateMatcher]="matcher">
+            <mat-error>
+              <span>Enter your message</span>
+            </mat-error>
+            <button type="submit" [disabled]="!chatForm.valid" mat-button matSuffix mat-icon-button aria-label="Send">
+              <mat-icon>send</mat-icon>
+            </button>
+          </mat-form-field>
+        </form>
+      </footer>
+    </mat-drawer-content>
+  </mat-drawer-container>
 </div>
