+++ src/app/components/chat-feed/chat-feed.component.ts
@@ -1,8 +1,30 @@
-import { Component, OnInit,OnChanges } from '@angular/core';
-import { ChatService } from 'src/app/services/chat/chat.service';
-import { Observable } from 'rxjs';
-import { AngularFireList } from '@angular/fire/database';
-import { Chatmessage } from 'src/app/classes/message/chatmessage';
+import { Component, OnInit, ElementRef, ViewChild, AfterViewChecked } from '@angular/core';
+import { Router, ActivatedRoute } from '@angular/router';
+import { FormControl, FormGroupDirective, FormBuilder, FormGroup, NgForm, Validators } from '@angular/forms';
+import { ErrorStateMatcher } from '@angular/material/core';
+import firebase from 'firebase/app';
+import 'firebase/auth'
+import { AngularFireAuth } from '@angular/fire/auth';
+import { AngularFireDatabase } from '@angular/fire/database';
+
+export class MyErrorStateMatcher implements ErrorStateMatcher {
+  isErrorState(control: FormControl | null, form: FormGroupDirective | NgForm | null): boolean {
+    const isSubmitted = form && form.submitted;
+    return !!(control && control.invalid && (control.dirty || control.touched || isSubmitted));
+  }
+}
+
+export const snapshotToArray = (snapshot: any) => {
+  const returnArr: any[] = [];
+
+  snapshot.forEach((childSnapshot: any) => {
+    const item = childSnapshot.val();
+    item.key = childSnapshot.key;
+    returnArr.push(item);
+  });
+
+  return returnArr;
+};
 
 
 @Component({
@@ -10,17 +32,98 @@ import { Chatmessage } from 'src/app/classes/message/chatmessage';
   templateUrl: './chat-feed.component.html',
   styleUrls: ['./chat-feed.component.css']
 })
-export class ChatFeedComponent implements OnInit,OnChanges {
+export class ChatFeedComponent implements OnInit {
+  @ViewChild('chatcontent') chatcontent: ElementRef;
+  scrolltop: any | null;
 
-  feed: Observable<any[]>;
+  chatForm: FormGroup;
+  chatname = '';
+  roomname = '';
+  message = '';
+  users: any[];
+  chats: any[];
+  user: any;
+  userName: any;
+  messages: any
+  matcher = new MyErrorStateMatcher();
 
-  constructor(private chat: ChatService) { }
+  constructor(private Auth: AngularFireAuth, private db: AngularFireDatabase, private router: Router,
+    private route: ActivatedRoute,
+    private formBuilder: FormBuilder) {
+    this.Auth.authState.subscribe(auth => {
+      if (auth !== undefined && auth !== null) {
+        this.user = auth;
+      }
+      this.getUser().valueChanges().subscribe(a => {
+        this.userName = a;
+        this.chatname = this.userName.displayName
+      });
 
-  ngOnInit(): void {
-    this.feed = this.chat.getMessages().valueChanges();
+      this.roomname = this.route.snapshot.params.roomname;
+      console.log(this.roomname)
+      firebase.database().ref('chats/').on('value', resp => {
+        let chats = snapshotToArray(resp);
+        this.chats = chats.filter(x => x.roomname === this.roomname)
+        setTimeout(() => this.scrolltop = this.chatcontent.nativeElement.scrollHeight, 500);
+      });
+      firebase.database().ref('roomusers/').orderByChild('roomname').equalTo(this.roomname).on('value', (resp2: any) => {
+        const roomusers = snapshotToArray(resp2);
+        this.users = roomusers.filter(x => x.status === 'online');
+      });
+    })
   }
 
-  ngOnChanges(): void {
-    this.feed = this.chat.getMessages().valueChanges();
+  getUser() {
+    const userId = this.user.uid;
+    const path = `/users/${userId}`;
+    return this.db.object(path);
+  }
+
+
+  ngOnInit(): void {
+    this.chatForm = this.formBuilder.group({
+      'message': [null, Validators.required]
+    });
+
+  }
+
+  onFormSubmit(form: any) {
+    var day = new Date();
+    const options = { day: 'numeric', month: 'long', year: "numeric", timeZone: "Asia/Kolkata" } as const;
+    const today = day.toLocaleDateString("en-IN", options);
+
+    const chat = form;
+    chat.roomname = this.roomname;
+    console.log(chat.roomname)
+    chat.chatname = this.chatname;
+    chat.date = today;
+    chat.type = 'message';
+    const newMessage = firebase.database().ref('chats/').push();
+    newMessage.set(chat);
+    this.chatForm = this.formBuilder.group({
+      'message': [null, Validators.required]
+    });
+  }
+  exitChat() {
+    const chat = { roomname: '', chatname: '', message: '', date: '', type: '' };
+    chat.roomname = this.roomname;
+    chat.chatname = this.chatname;
+    chat.date = new Date().toDateString();
+    chat.message = `${this.chatname} leave the room`;
+    chat.type = 'exit';
+    const newMessage = firebase.database().ref('chats/').push();
+    newMessage.set(chat);
+
+    firebase.database().ref('roomusers/').orderByChild('roomname').equalTo(this.roomname).on('value', (resp: any) => {
+      let roomuser = [];
+      roomuser = snapshotToArray(resp);
+      const user = roomuser.find(x => x.chatname === this.chatname);
+      if (user !== undefined) {
+        const userRef = firebase.database().ref('roomusers/' + user.key);
+        userRef.update({ status: 'offline' });
+      }
+    });
+
+    this.router.navigate(['/roomlist']);
   }
 }
