+++ src/app/components/chatroom/chatroom.component.html
@@ -3,8 +3,7 @@
    <app-navbar></app-navbar>
   </div>
   <div #scroller class="chatfeed">
-    <app-chat-feed></app-chat-feed>
-    <app-chat-form></app-chat-form>
+    <!-- <app-chat-feed></app-chat-feed> -->
   </div>
   <div class="row">
     <div class="col">
