+++ src/app/components/home/home-page/home-page.component.html
@@ -38,7 +38,7 @@
     <a
       mat-list-item
       class="nav-item nav-link"
-      routerLink="/chatroom"
+      routerLink="/chats"
       routerLinkActive="active"
       matBadge="8"
       matBadgePosition="after"
