+++ src/app/components/home/navbar/navbar.component.html
@@ -12,7 +12,7 @@
       
         <a mat-list-item href="" class="nav-item nav-link active" routerLink='/home' routerLinkActive='active'><mat-icon>home</mat-icon> Home</a>
         <a mat-list-item class="nav-item nav-link" routerLink='profile' routerLinkActive='active'><mat-icon>contact_page</mat-icon>Profile</a> 
-        <a mat-list-item class="nav-item nav-link" routerLink='/chatroom' routerLinkActive='active'  matBadge="8" matBadgePosition="before" matBadgeColor="warn"><mat-icon>forum</mat-icon>Chat</a>
+        <a mat-list-item class="nav-item nav-link" routerLink='/chats' routerLinkActive='active'  matBadge="8" matBadgePosition="before" matBadgeColor="warn"><mat-icon>forum</mat-icon>Chat</a>
         <a mat-list-item class="nav-item nav-link" routerLink='about' routerLinkActive='active'><mat-icon>contact_support</mat-icon>Help</a>  
         <a mat-list-item class="nav-item nav-link" *ngIf="(user|async)?.uid" (click)="logout()"> <mat-icon>logout</mat-icon>Logout</a>
